Refactored to use request-targets and make URIs standalone
`UriTargetInterface` was renamed to `UriInterface`, and removes the
ability to model non-URI request targets. This patch removes those
methods, and removes functionality related to non-URI targets.

`RequestInterface` adds the accessor `getRequestTarget()` which will return the
request-target; if not present in the instance, the request-target returned will
be an origin-form calculated from the composed URI (and the string "/" if no URI
is present or if the URI does not contain a path). It also adds the
mutator `withRequestTarget()`, which will create a new instance with the
request-target specified (and raise an exception if it contains any whitespace).