All mutator methods MUST return a new instance
Per current incarnation of the PSR-7 spec. We're debating whether or not to
change it. However, the liklihood of these situations occuring (`with*()`
methods not resulting in a change) plus the small computational overhead is not
large, so returning the same instance is at best a micro-optimization. As such,
this path now correctly returns clones from all `with*()` method permutations.