[#61] Split HTML and JSON factories to Response types
Per discussion started on #61, this pull request removes the
`StringResponse` factory class in favor of two dedicated response types,
one for each of HTML and JSON responses.

In order to achieve that goal, it introduces the
`InjectContentTypeTrait`, which will search for a Content-Type header in
the provided headers, and, if none is found, inject the provided
Content-Type.

This also allowed extracting the logic for encoding data to JSON to
another method.