[#91] Ensure withUri() overwrites Host header, regardless of case
As reported in #91, `withUri()` will fail to overwrite an existing
`Host` header if the case of the original header name does not exactly
match `Host`. This patch adds tests for all case permutations of the
header name, and adds logic to `withUri()` to remove an existing `Host`
header regardless of case if found.