Removes output buffer awareness from Server, emitters
By removing a call to `ob_start()` within `Server::listen()`, we can
solve the issue of detecting when we have both a response and content in
the output buffer; in most cases, we will have already sent headers,
which will cause an exception to be raised; we can also check the
_current_ output buffer and, if non-empty, raise an exception.

This means we can:

- Remove the `SapiEmitterTrait::flush()` implementation, and all calls
  to it.
- Remove the `$maxBufferLevel` argument to each emitter.
- Remove tests regarding interactions of emitters with the output buffer.

This is a backwards-incompatible change. However, it fixes a rather
tricky problem that occurs currently when mixing buffered output and
response instances.