Removes Content-Length injection from SAPI stream emitters
Per #251, this patch removes Content-Length injection from SAPI stream
emitters. It also renames the `checkForPreviousOutput()` method to
`assertNoPreviousOutput()` to make it more clear that the method is a
no-op under proper usage.

The `assertNoPreviousOutput()` method also borrows an idea from the
comment to the issue, indicating we can check for (1) a non-zero output
buffer, and (2) use `ob_get_length()` to determine if it has content.