Ensure that PHP doesn't override response code
If we don't pass the response status code to the function `header()` PHP
might silently change the response code, which is not really nice since
it should be up for the developers to define how they design the
application.

This silent change usually happens when one uses "Location" with a
status code that's not 201 or 3xx, and some people already discussed a
lot and essencially they've said that PHP will not modify its behaviour
regarding this because it allows high level code to modify (fix) this.

References:

- https://bugs.php.net/bug.php?id=70273
- https://bugs.php.net/bug.php?id=51749
- https://bugs.php.net/bug.php?id=74535