Consistency changes and expectation updates
`ServerRequestFactory` was incorrectly passing the body to the new
request constructor. The content being passed was string content, and
needs to be either a stream URI, resource, or `StreamInterface`
instance. Test was updated to demonstrate this, and the factory was
updated to create the stream.

A number of consistency updates were also made:

- License annotation in headers of affected files.
- Flow of mock expectations.