Add a trait for managing download responses
After doing a bit of searching, I didn't find anything that would help
with sending a downloaded response, instead of the default, which is to
send a text response. So I created this one, which sends six headers
that ensure that the response sent will be interpreted by the client as
a downloadable file.