revert: Reverts changes from #378
Post-release, users discovered problems with `UploadedFile` extending
`SplFileInfo`, particularly when a stream path is used to create the
instance. Additionally, there are some differences with how
`UploadedFileInterface` and `SplFileInfo` each define the return type
for `getSize()`; the former allows nullification, while the latter does
not, but allows raising an exception (something explicitly
disallowed in PSR-7).

This patch reverts the code and documentation changes associated with
the patch.