Refactor to be able to use stream or filename for response body
This is following on from what mwop suggested in that the body of the
response can be set from either a string or a file. It's making use of a
Stream object and the body is instantiated similarly to XmlResponse as
how that works makes a lot of sense in this case.